code/requirements.txt
pandas
pyarrow
fastparquet
huggingface-hub
matplotlib
numpy
datasets
Pillow
torch
transformers
faiss-cpu
